[fix] status bar UI bug fixed
