+++ frontend/app/src/main/res/values/dimens.xml
@@ -65,7 +65,7 @@
     <dimen name="bottom_sheet_text_size">18sp</dimen>
     <dimen name="feed_logo_width">80dp</dimen>
     <dimen name="feed_logo_height">40dp</dimen>
-    <dimen name="feed_toolbar_height">80dp</dimen>
+    <dimen name="feed_toolbar_height">100dp</dimen>
     <dimen name="post_padding">15dp</dimen>
     <dimen name="post_profile_image_width">40dp</dimen>
     <dimen name="post_profile_image_height">40dp</dimen>
